When selecting a date in floorplan form, or edit staff form, the date displayed is changed to the name of a special event if there is one

+++ FloorplanClassLibrary/SqliteDataAccess.cs
@@ -1945,7 +1945,7 @@ namespace FloorplanClassLibrary
                 if (result != null)
                 {
                     return new SpecialEventDate(
-                        result.Id,
+                        (int)result.Id,
                         DateOnly.Parse(result.DateOnly),
                         Enum.Parse<SpecialEventDate.OutlierType>(result.Type),
                         result.ShouldIgnoreSales == 0,
